Node API: Use factory pattern for apps - each app module must define a create_app() function - move get_app to core
